+++ rt_thread_3.1.3-ucosiii_3.03-wrapper/uCOS-III_Wrapper/samples/mutex_sample.c
@@ -16,9 +16,11 @@
 #define TASK_STACK_SIZE       256	/*任务堆栈大小*/
 #define TASK_TIMESLICE        5		/*任务时间片*/
 
+ALIGN(RT_ALIGN_SIZE)
 static CPU_STK AppTask1_Stack[TASK_STACK_SIZE];/*任务堆栈*/
 static OS_TCB  AppTask1_TCB;/*任务控制块*/
 
+ALIGN(RT_ALIGN_SIZE)
 static CPU_STK AppTask2_Stack[TASK_STACK_SIZE];/*任务堆栈*/
 static OS_TCB  AppTask2_TCB;/*任务控制块*/